requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
pytest==8.3.2
